Module 10 - Renamed py files to be less Trello specific.
